{{cookiecutter.project_slug}}/src/requirements.txt
cloudshell-shell-core>=2.0.0,<2.1.0
cloudshell-pdu-core>=1.0, <1.1